Version 1.3.5
Now uses ExpressJS to serve both the ReactJS content and the API endpoints.
Doing so means that the use of Concurrently to run both an express server and a react development server is unnecessary.

Uses process environment variables to define URL and PORT. Defaults are ' localhost' for URL and 3000 for PORT.